move getCommonMetadataHeadTags to test's utils
