+++ test/e2e/app-dir/metadata-files-static-output/metadata-files-static-output-group-route.test.ts
@@ -1,5 +1,5 @@
 import { nextTestSetup } from 'e2e-utils'
-import { getMetadataHeadTags } from 'next-test-utils'
+import { getCommonMetadataHeadTags } from './utils'
 import { getMetadataRouteSuffix } from 'next/dist/lib/metadata/get-metadata-route'
 
 describe('metadata-files-static-output-group-route', () => {
@@ -17,7 +17,7 @@ describe('metadata-files-static-output-group-route', () => {
   it('should have correct link tags for group page', async () => {
     const browser = await next.browser('/group')
 
-    expect(await getMetadataHeadTags(browser)).toMatchInlineSnapshot(`
+    expect(await getCommonMetadataHeadTags(browser)).toMatchInlineSnapshot(`
      {
        "links": [
          {
